+++ index.html
@@ -56,6 +56,6 @@
     </ul>
   </div>
 </section>
-<script src="js/app.js"></script>
+<script type="module" src="js/app.js"></script>
 </body>
 </html>
